requirements.txt
langchain-openai
python-dotenv == 1.0.1
langchain == 0.2.1
langchain-community == 0.2.1
langserve >= 0.0.51
langchain-anthropic == 0.1.15
langchain-google-genai == 1.0.5
langchain-google-firestore == 0.3.0
firestore == 0.0.8
chromadb == 0.5.0
tiktoken == 0.7.0
sentence-transformers == 3.0.0
bs4 == 0.0.2
firecrawl-py == 0.0.13
langchainhub == 0.1.18
wikipedia == 1.4.0
tavily-python == 0.3.3
fastapi
uvicorn
langserve
sse_starlette
httpx==0.23.3
streamlit